requirements.txt
# AI Policy Query System Dependencies - Python 3.13 Compatible

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.20.0

# Document Processing
pdfplumber>=0.10.0
python-docx>=0.8.11

# Machine Learning & Embeddings
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
numpy>=1.24.0

# API Integrations
openai>=1.0.0
requests>=2.31.0

# Utilities
python-multipart>=0.0.5
aiofiles>=23.0.0
typing-extensions>=4.5.0
psutil>=5.9.0
python-dotenv>=1.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 